Move sympy_commands to utils
